UPA2010 the UML Tobi Wooaahh it's finally plugintime. But it's bare ugly. See the email for more details. I've commented the points of interest via:
// Eingriffspunkt nummer x

where x is the number.. afaik there are 2. Ok I need a bit more since I gotta make sure that the magic only happens when the 2 elements of interest are selected but that shouldn't be too difficult. Up to now this really is just copy&paste.. the important parts will be added tomorrow... or to be more accurate later today.